+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- Two `UpnData` object with the same fields values considered as equals.
+
 ## 0.0.3 
 
 - Added `toString()` for `UpnData`.